Try moving LogicProgram to package.scala.
